requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
appnope==0.1.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
crayons==0.4.0
decorator==5.1.1
einops==0.4.0
executing==0.8.2
filelock==3.6.0
frozenlist==1.3.0
fsspec==2022.2.0
ftfy==6.1.1
future==0.18.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
huggingface-hub==0.4.0
idna==3.3
imageio==2.16.0
imageio-ffmpeg==0.4.5
imgtag==0.1.0
importlib-metadata==4.11.1
ipython==8.0.1
jedi==0.18.1
joblib==1.1.0
kornia==0.6.3
Markdown==3.3.6
matplotlib-inline==0.1.3
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.22.2
oauthlib==3.2.0
omegaconf==2.1.1
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
Pygments==2.11.2
pyparsing==3.0.7
python-xmp-toolkit==2.0.1
pytorch-lightning==1.5.10
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.47
six==1.16.0
stack-data==0.2.0
Stegano==0.9.7
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tokenizers==0.11.5
tomli==2.0.1
torch==1.10.2
torchmetrics==0.7.2
torchvision==0.11.3
tqdm==4.62.3
traitlets==5.1.1
transformers==4.16.2
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
wget==3.2
yarl==1.7.2
zipp==3.7.0
